requirements.txt
streamlit
pandas
langchain-experimental
langchain-google-genai
python-dotenv